Refactor FeatureParser test to follow cake pattern
