+++ src/main/scala/modelservice/Boot.scala
@@ -1,7 +1,8 @@
 package modelservice
 
 import modelservice.api.{Api, ModelServiceActor}
-import modelservice.core.{ModelServiceActors, InitCore}
+import modelservice.core.prediction.PredictionActorSet
+import modelservice.core.{CoreActorSet, InitCore}
 import modelservice.web.Web
 
-object Boot extends App with InitCore with Api with Web with ModelServiceActors
+object Boot extends App with InitCore with Api with Web with CoreActorSet with PredictionActorSet
