+++ src/main/scala/modelservice/api/Api.scala
@@ -7,7 +7,7 @@ import modelservice.core._
  * Api
  */
 trait Api {
-  this: Core with ActorSet =>
+  this: Core with CoreActors =>
 
   private implicit val _ = system.dispatcher
   val modelService = system actorOf Props(new ModelServiceActor(this))
