+++ src/main/scala/modelservice/core/FeatureParser.scala
@@ -5,14 +5,14 @@ import org.json4s.jackson.JsonMethods._
 import akka.actor._
 import akka.dispatch.{BoundedMessageQueueSemantics, RequiresMessageQueue}
 import spray.http.HttpEntity
-import modelservice.core.prediction.{PredictActor, TreePredictionActor}
+import modelservice.core.prediction.{PredictionActors, PredictActor, TreePredictionActor}
 
 /**
  * Actor to parse features
- * @param predictionActor Create actor using current the context.  This should be the prediction actor in
+ * @param predictionActors Create actor using current the context.  This should be the prediction actor in
  *                              production or a probe actor for tests.  Defaults to tree prediction
  */
-class FeatureParser(predictionActor: ActorRef)
+class FeatureParser(predictionActors: PredictionActors)
   extends Actor with ActorLogging with RequiresMessageQueue[BoundedMessageQueueSemantics] {
   import FeatureParser._
 //  import VectorizeFeatures._
@@ -24,7 +24,7 @@ class FeatureParser(predictionActor: ActorRef)
         // Start prediction tree
         val parsedContext = PredictTree(parse(record.asString).values.asInstanceOf[Map[String, Any]],
           modelKey, paramKey, modelStorage, client)
-        predictionActor ! parsedContext
+        predictionActors.predictionActor ! parsedContext
       } catch {
         case e: Exception => log.info(e.getLocalizedMessage)
       }
